made FS_WriteFile's len parameter size_t to fix some warnings
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5491 d7cf8633-e32d-0410-b094-e92efae38249

+++ fs.c
@@ -1934,7 +1934,7 @@ FS_WriteFile
 The filename will be prefixed by the current game directory
 ============
 */
-qboolean FS_WriteFile (const char *filename, void *data, int len)
+qboolean FS_WriteFile (const char *filename, void *data, size_t len)
 {
 	qfile_t *file;
